fix: glucose float

+++ ActiveDiabetesAssistant.PL.API/Models/GlucoseInfo/Queries/ViewModel.cs
@@ -2,6 +2,6 @@
 
 public class GlucoseInfoViewModel : BaseViewModel
 {
-	public int GlucoseData { get; set; }
+	public float GlucoseData { get; set; }
 	public int? StepsCount { get; set; }
 }
